Separate known_tx from init
